refactor(ivy): rename `styling_instructions` to `styling` (#29770)
PR Close #29770
