+++ packages/core/src/render3/instructions/all.ts
@@ -41,5 +41,5 @@ export * from './projection';
 export * from './property';
 export * from './property_interpolation';
 export * from './select';
-export * from './styling_instructions';
+export * from './styling';
 export * from './text';
